+++ api/v1/individualType/config.ini
