+++ api/v1/individualType/index.py
@@ -1,4 +1,3 @@
-# gunicorn index:app
 import falcon
 import RaceCard
 
@@ -7,4 +6,4 @@ class CORSMiddleware:
 		resp.set_header('Access-Control-Allow-Origin', '*')
 
 app = falcon.API(middleware=[CORSMiddleware()])
-app.add_route('/v1/indiv/raceCard', RaceCard.RaceCard())
+app.add_route('/v1/RaceCard', RaceCard.RaceCard())
