+++ api/v1/umabashira/Umabashira.py
@@ -1,0 +1,38 @@
+import json
+import requests
+import configparser
+import UmabashiraGenerator
+import RaceInfoGenerator
+
+class Umabashira:
+	def on_get(self, req, resp):
+		#設定ファイルの読み込み
+		inifile = configparser.ConfigParser()
+		inifile.read('./config.ini', 'UTF-8')
+		csvfile = inifile.get('jvdfile', 'csvfile_A')
+		txtfile = inifile.get('jvdfile', 'txtfile_D')
+
+		#レスポンスの作成
+		result = {}
+		result['result'] = {}
+		result['result']['type'] = "Umabashira"
+
+		#race_infoの取得
+		raceInfoGenerator = RaceInfoGenerator.RaceInfoGenerator(txtfile)
+		result['result'].update(raceInfoGenerator.getRaceInfo())
+
+		#umabashiraの取得
+		umabashiraGenerator = UmabashiraGenerator.UmabashiraGenerator(csvfile)
+		result['result']['horse'] = umabashiraGenerator.getUmabashira()
+
+		#typeが指定されるとき
+		if 'type' in req.params:
+			if req.params['type'] == "RaceCard":
+				for i in range(len(result['result']['horse'])):
+					jvdHorseId = result['result']['horse'][i]['org_horse_master_id_jvd']
+					url = "http://localhost:8070/v1/RaceCard?jvdHorseId={0}&size=4".format(jvdHorseId)
+					r = requests.get(url)
+					res = r.json()
+					result['result']['horse'][i]['item'] = res['result']
+
+		resp.body = json.dumps(result, ensure_ascii=False)
