+++ api/v1/umabashira/index.py
@@ -1,10 +1,9 @@
-# gunicorn index:app
 import falcon
-import Original
+import Umabashira
 
 class CORSMiddleware:
 	def process_request(self, req, resp):
 		resp.set_header('Access-Control-Allow-Origin', '*')
 
 app = falcon.API(middleware=[CORSMiddleware()])
-app.add_route('/v1/umabashira/original', Original.Original())
+app.add_route('/v1/Umabashira', Umabashira.Umabashira())
